Refactor layer model
